Airbnb listings and reviews

A sample dataset of Airbnb listings and reviews of properties from Sydney, Australia.

A sample dataset of Airbnb listings and reviews of properties from Sydney, Australia.

Usage

airbnb_listings

airbnb_reviews

airbnb_listings

airbnb_reviews

Format

airbnb_listings

A data.frame with 1623 rows and 68 columns

id

Unique identifier for the listing

name

Name of the listing

host_id

Unique identifier for the host

host_name

Name of the host

neighbourhood_group_cleansed

Neighbourhood group

neighbourhood_cleansed

Neighbourhood

latitude

Latitude of the listing

longitude

Longitude of the listing

room_type

Type of room

price

Price per night

minimum_nights

Minimum nights required

maximum_nights

Maximum nights allowed

availability_30

Number of days available in the next 30 days

availability_60

Number of days available in the next 60 days

availability_90

Number of days available in the next 90 days

availability_365

Number of days available in the next 365 days

number_of_reviews

Number of reviews

number_of_reviews_ltm

Number of reviews in the last 12 months

first_review

Date of the first review

last_review

Date of the last review

review_scores_rating

Rating score

review_scores_accuracy

Accuracy score

review_scores_cleanliness

Cleanliness score

review_scores_checkin

Check-in score

review_scores_communication

Communication score

review_scores_location

Location score

review_scores_value

Value score

instant_bookable

Instant booking available

calculated_host_listings_count

Number of listings by the host

calculated_host_listings_count_entire_homes

Number of entire homes by the host

calculated_host_listings_count_private_rooms

Number of private rooms by the host

calculated_host_listings_count_shared_rooms

Number of shared rooms by the host

reviews_per_month

Number of reviews per month

host_is_superhost

Host is a superhost

host_has_profile_pic

Host has a profile picture

host_identity_verified

Host identity is verified

host_response_time

Host response time

host_response_rate

Host response rate

host_acceptance_rate

Host acceptance rate

host_neighbourhood

Host neighbourhood

host_listings_count

Number of listings by the host

host_total_listings_count

Total number of listings by the host

host_verifications

Host verifications

host_has_profile_pic

Host has a profile picture

host_identity_verified

Host identity is verified

neighbourhood

Neighbourhood

neighbourhood_cleansed

Neighbourhood

neighbourhood_group_cleansed

Neighbourhood group

city

City

state

State

zipcode

Zipcode

market

Market

smart_location

Smart location

country_code

Country code

country

Country

latitude

Latitude

longitude

Longitude

is_location_exact

Location is exact

property_type

Property type

room_type

Room type

accommodates

Number of guests accommodated

bathrooms

Number of bathrooms

bedrooms

Number of bedrooms

beds

Number of beds

bed_type

Type of bed

amenities

Amenities

square_feet

Square feet

price

Price per night

weekly_price

Weekly price

monthly_price

Monthly price

security_deposit

Security deposit

cleaning_fee

Cleaning fee

s

airbnb_reviews

A data.frame with 5679 rows and 6 columns

listing_id

Unique identifier for the listing

id

Unique identifier for the review

date

Date of the review

reviewer_id

Unique identifier for the reviewer

reviewer_name

Name of the reviewer

comments

Text of the review

airbnb_listings

A data.frame with 1623 rows and 68 columns

id

Airbnb's unique identifier for the listing.

name

Name of the listing.

description

Detailed description of the listing.

neighborhood_overview

Host's description of the neighbourhood.

picture_url

URL to the Airbnb hosted regular sized image for the listing.

host_id

Airbnb's unique identifier for the host/user.

host_name

Name of the host. Usually just the first name(s).

host_since

The date the host/user was created. For hosts that are Airbnb guests this could be the date they registered as a guest.

host_location

The host's self reported location.

host_about

Description about the host.

host_response_time

The time interval between when a host responds to an inquiry from a guest.

host_response_rate

Percentage of inquiries from potential guests that are responded to by hosts.

host_acceptance_rate

That rate at which a host accepts booking requests.

host_is_superhost

Whether the host is a super host or not.

host_thumbnail_url

A thumbnail of the host.

host_picture_url

A URL to the picture of the host.

host_neighbourhood

The host neighbourhood.

host_listings_count

The number of listings the host has.

host_total_listings_count

The number of listings the host has.

host_verifications

Host communication verifications.

host_has_profile_pic

Whether the host has a profile pic.

host_identity_verified

Whether the host has their identity verified.

neighbourhood_cleansed

The neighbourhood as geocoded using the latitude and longitude against neighborhoods as defined by open or public digital shapefiles.

latitude

Uses the World Geodetic System (WGS84) projection for latitude and longitude.

longitude

Uses the World Geodetic System (WGS84) projection for latitude and longitude.

property_type

Self selected property type. Hotels and Bed and Breakfasts are described as such by their hosts in this field.

room_type

Entire home/apt, Private room, Shared room, or Hotel. Entire places are best if you're seeking a home away from home. With an entire place, you'll have the whole space to yourself. This usually includes a bedroom, a bathroom, a kitchen, and a separate, dedicated entrance. Hosts should note in the description if they'll be on the property or not (ex: "Host occupies first floor of the home"), and provide further details on the listing. Private rooms are great for when you prefer a little privacy, and still value a local connection. When you book a private room, you'll have your own private room for sleeping and may share some spaces with others. You might need to walk through indoor spaces that another host or guest may occupy to get to your room. Shared rooms are for when you don't mind sharing a space with others. When you book a shared room, you'll be sleeping in a space that is shared with others and share the entire space with other people. Shared rooms are popular among flexible travelers looking for new friends and budget-friendly stays.

accommodates

The maximum capacity of the listing.

bathrooms

The number of bathrooms in the listing.

bathrooms_text

The text of the number of bathsroom in the listings.

bedrooms

The number of bedrooms.

beds

The number of bed(s).

amenities

The amenities.

price

Daily price in local currency.

minimum_nights

Minimum number of night stay for the listing.

maximum_nights

Maximum number of night stay for the listing.

minimum_minimum_nights

The smallest minimum_night value from the calender (looking 365 nights in the future).

maximum_minimum_nights

The largest minimum_night value from the calender (looking 365 nights in the future).

minimum_maximum_nights

The smallest maximum_night value from the calender (looking 365 nights in the future).

maximum_maximum_nights

The largest maximum_night value from the calender (looking 365 nights in the future).

minimum_nights_avg_ntm

The average minimum_night value from the calender (looking 365 nights in the future).

maximum_nights_avg_ntm

The average maximum_night value from the calender (looking 365 nights in the future).

has_availability

Whether there is availability or not.

availability_30

The availability of the listing x days in the future as determined by the calendar. Note a listing may not be available because it has been booked by a guest or blocked by the host.

availability_60

availability_90

availability_365

number_of_reviews

The number of reviews the listing has.

number_of_reviews_ltm

The number of reviews the listing has (in the last 12 months).

number_of_reviews_l30d

The number of reviews the listing has (in the last 30 days).

first_review

The date of the first/oldest review.

last_review

The date of the last/newest review.

review_scores_rating

The review score for ratings of the listing.

review_scores_accuracy

The review score for accuracy of the listing.

review_scores_cleanliness

The review score for cleanliness of the listing.

review_scores_checkin

The review score for checkin experience of the listing.

review_scores_communication

The review score for communication of the listing.

review_scores_location

The review score for location of the listing.

review_scores_value

The review score for value of the listing.

license

The licence/permit/registration number.

instant_bookable

Whether the guest can automatically book the listing without the host requiring to accept their booking request. An indicator of a commercial listing.

calculated_host_listings_count

The number of listings the host has in the current scrape, in the city/region geography.

calculated_host_listings_count_entire_homes

The number of Entire home/apt listings the host has in the current scrape, in the city/region geography.

calculated_host_listings_count_private_rooms

The number of Private room listings the host has in the current scrape, in the city/region geography.

calculated_host_listings_count_shared_rooms

The number of Shared room listings the host has in the current scrape, in the city/region geography.

reviews_per_month

The average number of reviews per month the listing has over the lifetime of the listing.

airbnb_reviews

A data.frame with 5679 rows and 6 columns

listing_id

Unique identifier for the listing

id

Unique identifier for the review

date

Date of the review

reviewer_id

Unique identifier for the reviewer

reviewer_name

Name of the reviewer

comments

Text of the review

Source

https://insideairbnb.com/get-the-data/

https://insideairbnb.com/get-the-data/